feat: update theme and setup
- add an image shortcode and style figure elements
- correctly show all posts on posts page
- move slug filter function into separate file
- rename image files to include date
- use image shortcode in GitHub project post
- update inline code block styling
- add h4 style
