+++ src/scss/partials/footer.scss
@@ -1,5 +1,4 @@
 .footer {
-  border-top: 1px solid $link-colour--lighter;
   margin-top: 2rem;
 
   @supports (display: grid) {
