+++ src/scss/partials/main.scss
@@ -10,9 +10,9 @@ hr {
 }
 
 figure {
-  border: 10px solid #f2f2f2;
+  border: 1px solid #ddd;
   margin: 0 auto 1rem auto;
-  max-width: 620px;
+  max-width: 100%;
 }
 
 img {
