+++ src/scss/partials/syntax-highlighting.scss
@@ -29,7 +29,7 @@ pre[class*="language-"] {
 /* Code blocks */
 pre[class*="language-"] {
   padding: 1em;
-  margin: 0.5em 0;
+  margin: 0 0 1.25rem 0;
   overflow: auto;
   border-radius: 0.3em;
 }
