+++ src/scss/partials/typography.scss
@@ -47,6 +47,14 @@ h3 {
   margin-bottom: 1.125rem;
 }
 
+h4 {
+  @include font-size(1.125rem);
+
+  font-weight: 400;
+  margin: 0;
+  margin-bottom: 1rem;
+}
+
 ul,
 ol {
   line-height: 1.7;
@@ -82,13 +90,16 @@ pre {
 }
 
 code {
-  @include font-size(1rem);
-
-  background-color: #002b36;
+  background-color: desaturate($link-colour--lightest, 5%);
   border-radius: 4px;
-  color: #93a1a1;
-  padding: 0.25rem;
+  color: inherit;
+  padding: 0.15em 0.25em;
   word-break: break-word;
+
+  a & {
+    color: darken($link-colour, 15%);
+    font-weight: 500;
+  }
 }
 
 .feature {
