+++ src/scss/partials/variables.scss
@@ -8,3 +8,4 @@ $link-colour: slateblue;
 $link-colour--light: rgba(slateblue, 0.3);
 $link-colour--lighter: rgba(slateblue, 0.2);
 $link-colour--lightest: rgba(slateblue, 0.05);
+$link-colour--super-light: rgba(slateblue, 0.025);
